refactor: move _genai_config to integrations/ort_genai/ep_config
genai-related code belongs in the integrations directory alongside the
existing GenaiConfigGenerator. Moved:
  src/mobius/_genai_config.py → integrations/ort_genai/ep_config.py
  src/mobius/_genai_config_test.py → integrations/ort_genai/ep_config_test.py

Updated:
  integrations/ort_genai/__init__.py — exports make_provider_options,
    make_sliding_window_config, make_kv_cache_dim_name, make_genai_decoder_config

43 tests pass (20 ep_config + 23 ep_optimization).
