+++ src/mobius/integrations/ort_genai/__init__.py
@@ -7,8 +7,20 @@ All onnxruntime-genai specific code lives here. The core model/task/component
 layers remain runtime-agnostic.
 """
 
+from mobius.integrations.ort_genai.ep_config import (
+    make_genai_decoder_config,
+    make_kv_cache_dim_name,
+    make_provider_options,
+    make_sliding_window_config,
+)
 from mobius.integrations.ort_genai.genai_config import (
     GenaiConfigGenerator,
 )
 
-__all__ = ["GenaiConfigGenerator"]
+__all__ = [
+    "GenaiConfigGenerator",
+    "make_genai_decoder_config",
+    "make_kv_cache_dim_name",
+    "make_provider_options",
+    "make_sliding_window_config",
+]
